+++ NEWS.md
@@ -1,3 +1,5 @@
+# amt 0.1.6
+
 # amt 0.1.5
 ## updates
 - support for `hnorm` and `lnorm` step-length distributions.